Remove branch protection for quarkus-hibernate-search-extras (#154)
To enable auto-merge, we need to require status checks to pass before anyone can push to the branch. Which I did.

Problem is, this affects not only pull requests, but also direct pushes, which become impossible: one now has to push to a new branch, wait for status checks to pass, and only then can it get merged into `main`.

Since the Quarkiverse release process does not do that, it's fundamentally incompatible with requiring status checks to pass, and thus with auto-merge.

So, I'm giving up and reverting the configuration of `quarkus-hibernate-search-extras` to what it was before this whole experiment.

Sorry about the noise.